money_t changed to int64

+++ types.h
@@ -1,7 +1,7 @@
 #ifndef TYPES_H
 #define TYPES_H
 
-using money_t = double;
+using money_t = int64_t;
 using square_name_t = std::string;
 using player_name_t = std::string;
 using match_weight_t = double;
